requirements.txt
streamlit>=1.38.0
numpy>=1.26.0
langchain>=0.2.10
langchain-community>=0.2.10
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
torch>=2.1.0
transformers>=4.42.0
sentence-transformers>=3.0.0
